refactor: move step 1 of submit to own file
