+++ src/actions/submit/submit.ts
@@ -2,21 +2,11 @@ import graphiteCLIRoutes from '@withgraphite/graphite-cli-routes';
 import * as t from '@withgraphite/retype';
 import { request } from '@withgraphite/retyped-routes';
 import chalk from 'chalk';
-import prompts from 'prompts';
 import { API_SERVER } from '../../lib/api';
 import { execStateConfig } from '../../lib/config/exec_state_config';
 import { TContext } from '../../lib/context/context';
-import {
-  ExitFailedError,
-  KilledError,
-  PreconditionsFailedError,
-  ValidationFailedError,
-} from '../../lib/errors';
-import {
-  cliAuthPrecondition,
-  currentBranchPrecondition,
-} from '../../lib/preconditions';
-import { syncPRInfoForBranches } from '../../lib/sync/pr_info';
+import { ExitFailedError, PreconditionsFailedError } from '../../lib/errors';
+import { cliAuthPrecondition } from '../../lib/preconditions';
 import { getSurvey, showSurvey } from '../../lib/telemetry/survey/survey';
 import {
   detectUnsubmittedChanges,
@@ -26,19 +16,16 @@ import {
   logNewline,
   logSuccess,
   logTip,
-  logWarn,
 } from '../../lib/utils';
 import { assertUnreachable } from '../../lib/utils/assert_unreachable';
-import { isEmptyBranch } from '../../lib/utils/is_empty_branch';
 import { Unpacked } from '../../lib/utils/ts_helpers';
-import { MetaStackBuilder, Stack } from '../../wrapper-classes';
 import { Branch } from '../../wrapper-classes/branch';
 import { TScope } from '../scope';
-import { validateStack } from '../validate';
 import { getPRBody } from './pr_body';
 import { getPRDraftStatus } from './pr_draft';
 import { getPRTitle } from './pr_title';
 import { getReviewers } from './reviewers';
+import { getValidBranchesToSubmit } from './validate_branches';
 
 export type TSubmitScope = TScope | 'BRANCH';
 
@@ -85,7 +72,6 @@ export async function submitAction(
   },
   context: TContext
 ): Promise<void> {
-  let branchesToSubmit;
   // Check CLI pre-condition to warn early
   const cliAuthToken = cliAuthPrecondition(context);
   if (args.dryRun) {
@@ -106,25 +92,16 @@ export async function submitAction(
     args.draftToggle = true;
   }
 
-  // This supports the use case in sync.ts. Skips Steps 1 and 2
-  if (args.branchesToSubmit) {
-    branchesToSubmit = args.branchesToSubmit;
-  } else {
-    // Step 1: Validate
-    logInfo(
-      chalk.blueBright(
-        `✏️  [Step 1] Validating that this Graphite stack is ready to submit...`
-      )
-    );
-    const validationResult = await getValidBranchesToSubmit(
-      args.scope,
-      context
-    );
-    if (validationResult.result === 'ABORT') {
-      return;
-    }
-    branchesToSubmit = validationResult.submittableBranches;
+  // Step 1: Validate
+  // args.branchesToSubmit is for the sync flow. Skips Steps 1.
+  const branchesToSubmit =
+    args.branchesToSubmit ??
+    (await getValidBranchesToSubmit(args.scope, context));
+
+  if (!branchesToSubmit) {
+    return;
   }
+
   // Step 2: Prepare
   logInfo(
     chalk.blueBright(
@@ -182,162 +159,6 @@ export async function submitAction(
   }
 }
 
-async function getValidBranchesToSubmit(
-  scope: TSubmitScope,
-  context: TContext
-): Promise<
-  | {
-      result: 'SUCCESS';
-      submittableBranches: Branch[];
-    }
-  | { result: 'ABORT' }
-> {
-  let branchesToSubmit;
-  try {
-    if (scope === 'BRANCH') {
-      const currentBranch = currentBranchPrecondition(context);
-      branchesToSubmit = [currentBranch];
-    } else {
-      const stack = getStack(
-        {
-          currentBranch: currentBranchPrecondition(context),
-          scope: scope,
-        },
-        context
-      );
-      validateStack(scope, stack, context);
-      branchesToSubmit = stack.branches().filter((b) => !b.isTrunk(context));
-    }
-    logNewline();
-  } catch {
-    throw new ValidationFailedError(`Validation failed. Will not submit.`);
-  }
-
-  await syncPRInfoForBranches(branchesToSubmit, context);
-  const closedBranches = branchesToSubmit.filter(
-    (b) => b.getPRInfo()?.state === 'CLOSED'
-  );
-  const mergedBranches = branchesToSubmit.filter(
-    (b) => b.getPRInfo()?.state === 'MERGED'
-  );
-  const submittableBranches = branchesToSubmit.filter(
-    (b) =>
-      b.getPRInfo()?.state !== 'CLOSED' || b.getPRInfo()?.state !== 'MERGED'
-  );
-
-  const result =
-    processMergedBranches(mergedBranches) === 'ABORT' ||
-    processClosedBranches(closedBranches) === 'ABORT' ||
-    (await detectEmptyBranches(submittableBranches, context)) === 'ABORT'
-      ? 'ABORT'
-      : 'SUCCESS';
-
-  return { result, submittableBranches };
-}
-
-function processMergedBranches(mergedBranches: Branch[]): 'SUCCESS' | 'ABORT' {
-  if (mergedBranches.length === 0) {
-    return 'SUCCESS';
-  }
-
-  const hasMultipleBranches = mergedBranches.length > 1;
-
-  logError(
-    `PR${hasMultipleBranches ? 's' : ''} for the following branch${
-      hasMultipleBranches ? 'es have' : ' has'
-    } already been merged:`
-  );
-  mergedBranches.forEach((b) => logError(`▸ ${chalk.reset(b.name)}`));
-  logError(
-    `If this is expected, you can use 'gt repo sync' to delete ${
-      hasMultipleBranches ? 'these branches' : 'this branch'
-    } locally and restack dependencies.`
-  );
-
-  return 'ABORT';
-}
-
-function processClosedBranches(closedBranches: Branch[]): 'SUCCESS' | 'ABORT' {
-  if (closedBranches.length === 0) {
-    return 'SUCCESS';
-  }
-
-  const hasMultipleBranches = closedBranches.length > 1;
-
-  logError(
-    `PR${hasMultipleBranches ? 's' : ''} for the following branch${
-      hasMultipleBranches ? 'es have' : ' has'
-    } been closed:`
-  );
-  closedBranches.forEach((b) => logError(`▸ ${chalk.reset(b.name)}`));
-  logError(
-    `To submit ${
-      hasMultipleBranches ? 'these branches' : 'this branch'
-    }, please reopen the PR remotely.`
-  );
-
-  return 'ABORT';
-}
-
-export async function detectEmptyBranches(
-  submittableBranches: Branch[],
-  context: TContext
-): Promise<'SUCCESS' | 'ABORT'> {
-  const emptyBranches = submittableBranches.filter((branch) =>
-    isEmptyBranch(branch.name, getBranchBaseName(branch, context))
-  );
-  if (emptyBranches.length === 0) {
-    return 'SUCCESS';
-  }
-
-  const hasMultipleBranches = emptyBranches.length > 1;
-
-  logWarn(
-    `The following branch${
-      hasMultipleBranches ? 'es have' : ' has'
-    } no changes:`
-  );
-  emptyBranches.forEach((b) => logWarn(`▸ ${chalk.reset(b.name)}`));
-  logWarn(
-    `Are you sure you want to submit ${hasMultipleBranches ? 'them' : 'it'}?`
-  );
-  logNewline();
-
-  if (!execStateConfig.interactive()) {
-    return 'ABORT';
-  }
-
-  const response = await prompts(
-    {
-      type: 'select',
-      name: 'empty_branches_options',
-      message: `How would you like to proceed?`,
-      choices: [
-        {
-          title: `Abort command and keep working on ${
-            hasMultipleBranches ? 'these branches' : 'this branch'
-          }`,
-          value: 'fix_manually',
-        },
-        {
-          title: `Continue with empty branch${hasMultipleBranches ? 'es' : ''}`,
-          value: 'continue_empty',
-        },
-      ],
-    },
-    {
-      onCancel: () => {
-        throw new KilledError();
-      },
-    }
-  );
-  logNewline();
-
-  return response.empty_branches_options === 'continue_empty'
-    ? 'SUCCESS'
-    : 'ABORT';
-}
-
 async function submitPullRequests(
   args: {
     submissionInfoWithBranches: (Unpacked<TPRSubmissionInfo> & {
@@ -362,29 +183,6 @@ async function submitPullRequests(
 
   saveBranchPRInfo(prInfo, context);
   printSubmittedPRInfo(prInfo);
-}
-
-function getStack(
-  args: { currentBranch: Branch; scope: TScope },
-  context: TContext
-): Stack {
-  switch (args.scope) {
-    case 'UPSTACK':
-      return new MetaStackBuilder().upstackInclusiveFromBranchWithParents(
-        args.currentBranch,
-        context
-      );
-    case 'DOWNSTACK':
-      return new MetaStackBuilder().downstackFromBranch(
-        args.currentBranch,
-        context
-      );
-    case 'FULLSTACK':
-      return new MetaStackBuilder().fullStackFromBranch(
-        args.currentBranch,
-        context
-      );
-  }
 }
 
 /**
@@ -587,16 +385,6 @@ async function requestServerToSubmitPRs(
   } catch (error) {
     throw new ExitFailedError(`Failed to submit PRs`, error);
   }
-}
-
-function getBranchBaseName(branch: Branch, context: TContext): string {
-  const parent = branch.getParentFromMeta(context);
-  if (parent === undefined) {
-    throw new PreconditionsFailedError(
-      `Could not find parent for branch ${branch.name} to submit PR against. Please checkout ${branch.name} and run \`gt upstack onto <parent_branch>\` to set its parent.`
-    );
-  }
-  return parent.name;
 }
 
 async function pushMetaStacks(branchesPushedToRemote: Branch[]): Promise<void> {
